fix changes consumers

+++ src/CompanyService.Broker/Consumers/ChangeUserOfficeConsumer.cs
@@ -1,6 +1,7 @@
 ﻿using LT.DigitalOffice.CompanyService.Data.Interfaces;
 using LT.DigitalOffice.CompanyService.Mappers.Db.Interfaces;
 using LT.DigitalOffice.Kernel.Broker;
+using LT.DigitalOffice.Kernel.Exceptions.Models;
 using LT.DigitalOffice.Models.Broker.Requests.Company;
 using MassTransit;
 using System.Threading.Tasks;
@@ -9,11 +10,17 @@ namespace LT.DigitalOffice.CompanyService.Broker.Consumers
 {
     public class ChangeUserOfficeConsumer : IConsumer<IChangeUserOfficeRequest>
     {
+        private readonly IOfficeRepository _officeRepository;
         private readonly IOfficeUserRepository _officeUserRepository;
         private readonly IDbOfficeUserMapper _mapper;
 
         private object ChangeUserOffice(IChangeUserOfficeRequest request)
         {
+            if (!_officeRepository.Contains(request.OfficeId))
+            {
+                throw new BadRequestException($"No office with Id {request.OfficeId}.");
+            }
+
             _officeUserRepository.Remove(request.UserId);
             _officeUserRepository.Add(_mapper.Map(request));
 
@@ -21,9 +28,11 @@ namespace LT.DigitalOffice.CompanyService.Broker.Consumers
         }
 
         public ChangeUserOfficeConsumer(
+            IOfficeRepository officeRepository,
             IOfficeUserRepository officeUserRepository,
             IDbOfficeUserMapper mapper)
         {
+            _officeRepository = officeRepository;
             _officeUserRepository = officeUserRepository;
             _mapper = mapper;
         }
